Picky picky CVTexture
CVTexture is not so much an OpenCV object, it's a spin on functionality
that is used often in Unreal projects involving a dynamically generated
textures. Unfortunately this functionality depends on an asynchronous
call to update texture data, and I've yet to figure out why my
implementation inconsistently crashes unreal in some situations.